Graph and Subgraph new architecture
On this commit, I've realised a few upgrades I should do before start
inserting a evaluation subsystem.

Now, the architecture is more deep, by transforming the retrieval node into
a subgraph component, which could be used for many tasks (not only for the main query).

Also, it has been added a question_decomposition and qa_context, based on a few studies and
a medium article. Although it generates more token consumption, it improves the context and
the final answer (or at least this is the promisse, as I didn't evaluated it with data, but only with a few examples).

Finally, I've updated the README.md and divided the graph.py python file by separating the graph builder from the graph compiled.