Changed the event transmission to avoid spam
Before this commit, the API was sending all events, including the ones from LangGraph subgraphs.
This was causing a lot of spam in the frontend.

Now, the API only sends the events from the main workflow, and not from the LangGraph subgraphs.
Also, there is a new function in the schemas.py file that is responsible for avoiding spam, and all schemas class have a 'tags' field for this purpose.